Move StackStack to its own module
