+++ hcpl_fft/src/lib.rs
@@ -1,11 +1,13 @@
 #[cfg(feature = "real")]
 pub mod real;
 
+mod stack_stack;
+
 use hcpl_algebra::monoid::MultiplicativeIdentity;
 use hcpl_number_theory::roots::TryNthRootOfUnity;
 use std::fmt::Debug;
 
-/// Performs the bit-reversal permutation on `a`
+/// Performs the bit-reversal permutation on `vec`
 pub fn bit_reversal<T>(vec: &mut [T]) {
     let n = vec.len();
 
@@ -24,52 +26,29 @@ pub fn bit_reversal<T>(vec: &mut [T]) {
     }
 }
 
-mod stack_stack {
-    use std::mem::MaybeUninit;
+fn get_roots<T>(mut n: usize, inv: bool) -> stack_stack::StackStack<T>
+where
+    T: TryNthRootOfUnity,
+    <T as TryNthRootOfUnity>::Error: Debug,
+{
+    let mut roots = stack_stack::StackStack::new();
 
-    const BUF_SIZE: usize = 100;
-
-    /// A stack-allocated stack
-    pub(crate) struct StackStack<T> {
-        buffer: [MaybeUninit<T>; BUF_SIZE],
-        len: usize,
+    let mut last: T = if inv {
+        TryNthRootOfUnity::try_nth_root_of_unity_inv(n)
+    } else {
+        TryNthRootOfUnity::try_nth_root_of_unity(n)
+    }
+    .unwrap();
+    while n >= 2 {
+        roots.push(last);
+        n /= 2;
+        last = last * last;
     }
 
-    impl<T> StackStack<T> {
-        pub(crate) fn new() -> Self {
-            Self {
-                buffer: unsafe {
-                    MaybeUninit::<[MaybeUninit<T>; BUF_SIZE]>::uninit().assume_init()
-                },
-                len: 0,
-            }
-        }
-
-        pub(crate) fn push(&mut self, val: T) -> Option<()> {
-            if self.len == BUF_SIZE {
-                return None;
-            }
-
-            self.buffer[self.len].write(val);
-            self.len += 1;
-
-            Some(())
-        }
-
-        pub(crate) fn pop(&mut self) -> Option<T> {
-            if self.len == 0 {
-                return None;
-            }
-
-            self.len -= 1;
-
-            // SAFETY: this element will have been written to exactly once, and won't be read again
-            Some(unsafe { self.buffer[self.len].assume_init_read() })
-        }
-    }
+    roots
 }
 
-/// Performs the in-place Fast Fourier Transform on the slice `a`, whose lenght must be a power of two. DOES NOT
+/// Performs the in-place Fast Fourier Transform on the slice `vec`, whose lenght must be a power of two. DOES NOT
 /// PERFORM NORMALISATION
 pub fn fft<T>(vec: &mut [T], inv: bool)
 where
@@ -80,21 +59,7 @@ where
 
     let n = vec.len();
 
-    let mut roots = stack_stack::StackStack::new();
-
-    let mut q = n;
-    let mut last: T = if inv {
-        TryNthRootOfUnity::try_nth_root_of_unity_inv(n)
-    } else {
-        TryNthRootOfUnity::try_nth_root_of_unity(n)
-    }
-    .unwrap();
-    while q >= 2 {
-        roots.push(last);
-        q /= 2;
-        last = last * last;
-    }
-
+    let mut roots = get_roots(n, inv);
     let mut width = 2;
     while width <= n {
         let w_d = roots.pop().unwrap();
